csvFinder-master/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dialogflow==1.0.0
Flask==1.1.2
future==0.18.2
fuzzywuzzy==0.17.0
google-api-core==1.17.0
google-auth==1.16.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-Levenshtein-wheels==0.13.1
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
